refactor: 공통 예외 처리 및 도메인 용어 관리 개선
- 기존 공통적인 예외를 통합
- 도메인에서 부분적으로 사용되던 domainName을 DomainTerm ENUM으로 대체하여 코드 간결화
- Theme 관련 쿼리에서 존재 여부를 확인하는 existsById, existsByName 메서드 추가

+++ src/main/java/roomescape/theme/application/usecase/ThemeCommandUseCaseImpl.java
@@ -3,9 +3,13 @@ package roomescape.theme.application.usecase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.common.domain.DomainTerm;
+import roomescape.common.exception.DuplicateException;
+import roomescape.common.exception.NotFoundException;
 import roomescape.theme.application.dto.CreateThemeServiceRequest;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeId;
+import roomescape.theme.domain.ThemeName;
 import roomescape.theme.domain.ThemeRepository;
 
 @RequiredArgsConstructor
@@ -14,15 +18,26 @@ import roomescape.theme.domain.ThemeRepository;
 public class ThemeCommandUseCaseImpl implements ThemeCommandUseCase {
 
     private final ThemeRepository themeRepository;
+    private final ThemeQueryUseCase themeQueryUseCase;
 
     @Override
     public Theme create(final CreateThemeServiceRequest request) {
-        return themeRepository.save(
-                request.toDomain());
+        final Theme theme = request.toDomain();
+        final ThemeName name = theme.getName();
+
+        if (themeQueryUseCase.existsByName(name)) {
+            throw new DuplicateException(DomainTerm.THEME, name);
+        }
+
+        return themeRepository.save(theme);
     }
 
     @Override
     public void delete(final ThemeId id) {
-        themeRepository.deleteById(id);
+        if (themeQueryUseCase.existsById(id)) {
+            themeRepository.deleteById(id);
+            return;
+        }
+        throw new NotFoundException(DomainTerm.THEME, id);
     }
 }
